+++ CHANGES.md
@@ -3,6 +3,7 @@ NEXT
 - drf : improve missing introspection endpoint error handling
 - have all `django_pyoidc` exceptions inherit from `DjangoOIDCException`
 - improve documentation for drf integration : `use_introspection_on_access_tokens` and `hook_validate_access_token`
+- add `hook_session_logout` setting #54 to allow the configuration of backchannel logout behaviour
 
 1.0.8
 -----